cardBrands test moved to separate files inside a 'cardBrands' directory
